IP-Adapter training pipeline, production-ready for M4 Pro 48GB.

Highlights:
- Manifest-driven ProductDataset with proper train/val split
- In-loop _validate() with deterministic-seed noise sampling so val
  loss is comparable across steps; file logging to train.log
- Gradient checkpointing wired through so the configured flag is
  actually honored (was a dead knob)
- MPS-compatible fp32 stack — autocast fp16/bf16 proved incompatible
  with SDXL+adapter on Apple Silicon after extensive smoke testing;
  all from_pretrained calls pin torch_dtype=torch.float32
- 512-res training to fit M4 Pro 48GB in a ~9h/platform budget
- Launcher PYTHONPATH fix so accelerate-launched scripts can import
  the adapters package (workaround for pyproject.toml broken
  build-backend)
- 10 unit tests for ProductDataset (happy path, both filename
  resolution forms, four error paths)
- Teammate handoff guide at docs/apple-silicon-training-handoff.md
- LoRA call site updated to new ProductDataset signature

Full Shopify 3000-step training (9h) completed with healthy
converging val loss: 0.073747 → 0.072500, monotonic decrease through
step 2000 then clean plateau. Etsy and eBay runs pending.

+++ adapters/ip_adapter/model.py
@@ -136,8 +136,8 @@ class IPAttnProcessor2_0(nn.Module):
 
         # IP-Adapter image cross-attention — only when embeddings are provided
         if ip_hidden_states is not None:
-            ip_key   = self.to_k_ip(ip_hidden_states)
-            ip_value = self.to_v_ip(ip_hidden_states)
+            ip_key   = self.to_k_ip(ip_hidden_states).to(query.dtype)
+            ip_value = self.to_v_ip(ip_hidden_states).to(query.dtype)
             ip_key   = ip_key.view(batch_size, -1, attn.heads, head_dim).transpose(1, 2)
             ip_value = ip_value.view(batch_size, -1, attn.heads, head_dim).transpose(1, 2)
             ip_out = F.scaled_dot_product_attention(
@@ -181,7 +181,9 @@ class IPAdapterSDXL(nn.Module):
         self.adapter_scale = adapter_scale
         self.num_tokens = num_tokens
 
-        self.image_encoder = CLIPVisionModelWithProjection.from_pretrained(image_encoder_id)
+        self.image_encoder = CLIPVisionModelWithProjection.from_pretrained(
+            image_encoder_id, torch_dtype=torch.float32,
+        )
         self.feature_extractor = CLIPImageProcessor.from_pretrained(image_encoder_id)
 
         clip_embed_dim = self.image_encoder.config.projection_dim
